Extract tag-out commands our of CombatCommand. Add one more user case scenario of removing tag-out
